requirements.txt

dash
plotly
pandas
pymongo
psycopg2-binary
python-dotenv
requests
riotwatcher
